Rename InputField to DelayedInput
